move registry to homescreen, add balances
