+++ modals/sendModal.html
@@ -22,6 +22,7 @@
     padding: 10px;
     border-radius: 15px;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
+    margin-bottom: -30px;
   }
 
   .radio-group input[type='radio'] {
@@ -72,7 +73,7 @@
               <label for="option2"> private </label>
             </form>
             <br>
-            <div class="form-check"id="withdrawalDiv" style="display: none;">
+            <div class="form-check"id="withdrawalDiv" style="display: none; margin-top:25px; margin-bottom:-15px;">
               <input class="form-check-input" type="checkbox" id="withdrawalCheckbox">
               <label class="form-check-label" for="withdrawalCheckbox">
                 Withdrawal
@@ -91,7 +92,7 @@
                                   <button class="btn btn-outline-primary btn-max" type="button" id="maxButton">MAX</button>
                                 </div>
                               </div>                              
-                              <button class="bridge-down w-100" type="button" id="signAndSend" style="margin-top:0px;">send</button>
+                              <button class="bridge-down w-100" type="button" id="signAndSend" style="margin-top:0px;">send chd</button>
                               <p class="text-muted"></p>
                           </form>
                       </div>
