+++ pages/amm.html
@@ -41,6 +41,9 @@
               <a class="nav-link" href="../pages/index.html">home</a>
             </li>
             <li class="nav-item">
+              <a class="nav-link" href="../pages/chd.html">chd</a>
+            </li>
+            <li class="nav-item">
               <a class="nav-link active" href="../pages/amm.html">trade</a>
             </li>
             <li class="nav-item">
@@ -48,9 +51,6 @@
             </li>
             <li class="nav-item">
               <a class="nav-link" href="../pages/cit.html">cit</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="../pages/registery.html">registry</a>
             </li>
           </ul>
         </div>
@@ -69,18 +69,18 @@
         <div class="row">
           <div class="col-md-6">
             <h2 class="my-4" style="font-size: 24px; font-weight: 700">
-              my balances
+              your balances
             </h2>
             <div class="table-responsive">
-              <table class="table">
+              <table class="table table-dark">
                 <thead class="text-center">
                   <tr>
-                    <th class="text-center placeholder" id="firstRow"></th>
+                    <th class="text-center placeholder" id="firstRow">...</th>
                   </tr>
                 </thead>
                 <tbody>
                   <tr>
-                    <td class="text-start" id="firstBase"></td>
+                    <td class="text-start" id="firstBase">...</td>
                     <td class="text-center" id="firstBal">...</td>
                   </tr>
                   <tr class="text-center">
@@ -95,15 +95,15 @@
               </table>
             </div>
             <div class="table-responsive">
-              <table class="table">
+              <table class="table table-dark">
                 <thead class="text-center">
                   <tr>
-                    <th class="text-center placeholder" id="secondRow"></th>
+                    <th class="text-center placeholder" id="secondRow">...</th>
                   </tr>
                 </thead>
                 <tbody>
                   <tr>
-                    <td class="text-start" id="secondBase"></td>
+                    <td class="text-start" id="secondBase">...</td>
                     <td class="text-center" id="secondBal">...</td>
                   </tr>
                   <tr class="text-center">
@@ -118,15 +118,15 @@
               </table>
             </div>
             <div class="table-responsive">
-              <table class="table">
+              <table class="table table-dark">
                 <thead class="text-center">
                   <tr>
-                    <th class="text-center placeholder" id="thirdRow"></th>
+                    <th class="text-center placeholder" id="thirdRow">...</th>
                   </tr>
                 </thead>
                 <tbody>
                   <tr>
-                    <td class="text-start" id="thirdBase"></td>
+                    <td class="text-start" id="thirdBase">...</td>
                     <td class="text-center" id="thirdBal">...</td>
                   </tr>
                   <tr class="text-center">
