+++ pages/auction.html
@@ -40,6 +40,9 @@
               <a class="nav-link" href="../pages/index.html">home</a>
             </li>
             <li class="nav-item">
+              <a class="nav-link" href="../pages/chd.html">chd</a>
+            </li>
+            <li class="nav-item">
               <a class="nav-link" href="../pages/amm.html">trade</a>
             </li>
             <li class="nav-item">
@@ -49,9 +52,6 @@
             </li>
             <li class="nav-item">
               <a class="nav-link" href="../pages/cit.html">cit</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="../pages/registery.html">registry</a>
             </li>
           </ul>
         </div>
