+++ pages/cit.html
@@ -17,12 +17,12 @@
             <div class="collapse navbar-collapse" id="navcol-1">
                 <ul class="navbar-nav mx-auto">
                     <li class="nav-item"><a class="nav-link" href="../pages/index.html">home</a></li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="../pages/chd.html">chd</a>
+                      </li>
                     <li class="nav-item"><a class="nav-link" href="../pages/amm.html">trade</a></li>
                     <li class="nav-item"><a class="nav-link" href="../pages/auction.html">auction</a></li>
                     <li class="nav-item"><a class="nav-link active" href="../pages/cit.html">cit</a></li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="../pages/registery.html">registry</a>
-                      </li>
                 </ul>
             </div>
         </div>
